reorganize routes using blueprints
