+++ templates/base.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
     <!-- Custom CSS -->
     <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
     <title>{% block title %}Project{% endblock %}</title>
@@ -13,14 +13,19 @@
     <!-- Nav bar -->
     <nav class="navbar navbar-dark bg-dark">
         <div class="container-fluid">
-            <a class="navbar-brand" href="{{ url_for('index') }}">[APP NAME]</a>
+            <a href="{{ url_for('index') }}" class="navbar-brand">[APP NAME]</a>
             <div class="d-flex">
                 {% if current_user.is_authenticated %}
-                    <a href="{{ url_for('customer_dashboard') }}" class="nav-link text-white me-2">Dashboard</a>
-                    <a href="{{ url_for('logout') }}" class="nav-link text-white">Logout</a>
+                    {% if current_user.user_type == 'vendor' %}
+                        <a href="{{ url_for('vendor.dashboard') }}" class="nav-link text-white me-2">Dashboard</a>
+                        <a href="{{ url_for('vendor.menu') }}" class="nav-link text-white me-2">Menu</a>
+                    {% else %}
+                        <a href="{{ url_for('customer.dashboard') }}" class="nav-link text-white me-2">Dashboard</a>
+                    {% endif %}
+                    <a href="{{ url_for('auth.logout') }}" class="nav-link text-white">Logout</a>
                 {% else %}
-                    <a href="{{ url_for('login') }}" class="nav-link text-white me-2">Login</a>
-                    <a href="{{ url_for('signup') }}" class="nav-link text-white">Sign Up</a>
+                    <a href="{{ url_for('auth.signup') }}" class="nav-link text-white me-2">Sign Up</a>
+                    <a href="{{ url_for('auth.login') }}" class="nav-link text-white">Login</a>
                 {% endif %}
             </div>
         </div>
