+++ templates/login.html
@@ -8,7 +8,7 @@
                 <h3>Login</h3>
             </div>
             <div class="card-body">
-                <form method="POST" action="{{ url_for('login') }}">
+                <form method="POST" action="{{ url_for('auth.login') }}">
                     <div class="mb-3">
                         <label for="email" class="form-label">Email</label>
                         <input type="email" class="form-control" id="email" name="email" required>
@@ -21,7 +21,7 @@
                 </form>
                 <hr>
                 <p class="text-center mb-0">
-                    Don't have an account? <a href="{{ url_for('signup') }}">Sign up</a>
+                    Don't have an account? <a href="{{ url_for('auth.signup') }}">Sign up</a>
                 </p>
             </div>
         </div>
