+++ templates/signup.html
@@ -8,7 +8,7 @@
                 <h3>Sign Up</h3>
             </div>
             <div class="card-body">
-                <form method="POST" action="{{ url_for('signup') }}" id="signupForm">
+                <form method="POST" action="{{ url_for('auth.signup') }}" id="signupForm">
                     <div class="mb-3">
                         <label for="username" class="form-label">Username</label>
                         <input type="text" class="form-control" id="username" name="username" required>
@@ -40,7 +40,7 @@
                 </form>
                 <hr>
                 <p class="text-center mb-0">
-                    Already have an account? <a href="{{ url_for('login') }}">Login</a>
+                    Already have an account? <a href="{{ url_for('auth.login') }}">Login</a>
                 </p>
             </div>
         </div>
